refactor(benchmark): unify shared helper utilities
