+++ package.json
@@ -41,8 +41,8 @@
     "benchmark:sync:upload:local": "node scripts/run-sync-bench.mjs local prime",
     "benchmark:sync:upload:remote": "node scripts/run-sync-bench.mjs remote prime",
     "benchmark:sync:remote": "node scripts/run-sync-bench.mjs remote",
-    "benchmark:playground:live-write": "node scripts/bench-playground-live-write.mjs",
-    "benchmark:sync:bootstrap": "node scripts/bench-discovery-connect.mjs",
+    "benchmark:playground:live-write": "pnpm -C packages/treecrdt-benchmark run build && node scripts/bench-playground-live-write.mjs",
+    "benchmark:sync:bootstrap": "pnpm -C packages/treecrdt-benchmark run build && node scripts/bench-discovery-connect.mjs",
     "benchmark:wasm": "pnpm -C packages/treecrdt-wasm-js run benchmark",
     "benchmark:web": "pnpm -C packages/treecrdt-wa-sqlite/e2e run bench",
     "test": "pnpm run test:browser && pnpm run test:native-node",
